+++ remote/servers.go
@@ -123,6 +123,15 @@ func (c *client) SetTransferStatus(ctx context.Context, uuid string, successful 
 	return nil
 }
 
+func (c *client) SetImportStatus(ctx context.Context, uuid string, data ImportStatusRequest) error {
+	resp, err := c.Post(ctx, fmt.Sprintf("/servers/%s/import", uuid), data)
+	if err != nil {
+		return err
+	}
+	_ = resp.Body.Close()
+	return nil
+}
+
 // ValidateSftpCredentials makes a request to determine if the username and
 // password combination provided is associated with a valid server on the instance
 // using the Panel's authentication control mechanisms. This will get itself